fastapi_app/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.128.0
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
requests==2.32.5
setuptools==80.9.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0
wheel==0.45.1
psycopg2-binary  # DB 접속용
sqlalchemy       # ORM
sentence-transformers  # ★ 핵심: 텍스트를 벡터로 바꿔주는 AI 모델
pgvector         # DB 벡터 연산 지원
httpx            # 비동기 HTTP 클라이언트 (외부 API 호출용)
python-dotenv    # 환경 변수 관리
